feat: add base DTO for every Nook entities, and write some functions for service

+++ src/main/java/com/library/entity/Booking.java
@@ -2,7 +2,6 @@ package com.library.entity;
 
 import com.library.enums.Status;
 import jakarta.persistence.*;
-import jdk.jfr.Timestamp;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -10,6 +9,7 @@ import lombok.Setter;
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
@@ -25,7 +25,7 @@ public class Booking {
     private LocalDateTime checkInDate;
     private LocalDateTime checkOutDate;
 
-    private Integer totalPrice;
+    private BigDecimal totalPrice;
 
     @Enumerated(EnumType.STRING)
     private Status status;
