+++ src/main/java/com/library/entity/Listing.java
@@ -5,6 +5,7 @@ import lombok.*;
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.UUID;
@@ -21,7 +22,7 @@ public class Listing {
 
     private String description;
 
-    private Integer pricePerNight;
+    private BigDecimal pricePerNight;
 
     @CreationTimestamp
     private LocalDateTime createdAt;
